Refactor to multiple components
